bisect.jpl: LABS_WHITELIST and TREES_WHITELIST parameters
Add LABS_WHITELIST and TREES_WHITELIST job parameters to only run
bisections if the lab name or kernel tree are on those whitelists.  If
the filter fails, the job is aborted.  When these parameters are not
defined, no filtering is applied and all jobs are run.  Labs and trees
filters are independent with labs being filtered first.

This is useful for bisections automatically triggered on all
regressions, which is the usual case in kernelci-backend.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>